Maintenance: Moves OAuth token management
Moves the OAuth token management functionality from a generic `auth` directory to a `base` directory within the `clients` directory.

This change improves the project's structure by aligning the token management more closely with the HTTP client infrastructure, as it's integral to the base client authentication process.

Updates relevant imports and references to reflect the new location.
