+++ src/index.ts
@@ -13,7 +13,7 @@ export { SFCCDevServer } from './core/server.js';
 export { SFCCLogClient } from './clients/log-client.js';
 export { SFCCDocumentationClient } from './clients/docs-client.js';
 export { OCAPIClient } from './clients/ocapi-client.js';
-export { TokenManager } from './auth/oauth-token.js';
+export { TokenManager } from './clients/base/oauth-token.js';
 export type { SFCCConfig, LogLevel, LogFileInfo, LogSummary, OCAPIConfig, OAuthToken, OAuthTokenResponse } from './types/types.js';
 
 // For direct execution, delegate to main.ts
